Update types and import
