--- .env.sample
@@ -1,5 +1,0 @@
-MAIL_HOST=your-mail-host
-MAIL_USER=your-mail-user
-MAIL_PASSWORD=your-mail-password
-MAIL_FROM=your-mail-from
-MAIL_ENCRYPTION=your-mail-encryption
