--- docker-compose.yml
@@ -1,76 +1,0 @@
-version: "2"
-#if needed to map the orthancPACS to a different hard disk device (after mounting it in path)
-# volumes: ["/home/petctviewer/orthanc:/var/lib/orthanc/db:Z"]
-
-services:
-    
-    gaelo-flow:
-      container_name: gaelo-flow
-      image: ghcr.io/pixilib/gaelo-flow-backend
-      depends_on: [redis, flow-db, orthanc-flow]
-      ports: ["3000:3000"]
-      environment:
-        APP_URL : "http://localhost"
-        ORTHANC_ADDRESS : "http://orthanc-flow:8042"
-        ORTHANC_USERNAME: "orthanc"
-        ORTHANC_PASSWORD: "orthanc"
-        TYPEORM_TYPE : "postgres"
-        TYPEORM_HOST : "flow-db"
-        TYPEORM_PORT : 5432
-        TYPEORM_USERNAME : "postgres"
-        TYPEORM_PASSWORD : "postgres"
-        TYPEORM_DATABASE : "gaelo-flow"
-        REDIS_ADDRESS : "redis"
-        REDIS_PORT : 6379
-        API_PORT : 3000
-        JWT_SECRET : "verySecretKey"
-        MAIL_HOST: "${MAIL_HOST}"
-        MAIL_USER: "${MAIL_USER}"
-        MAIL_PASSWORD: "${MAIL_PASSWORD}"
-        MAIL_FROM: "${MAIL_FROM}"
-        MAIL_ENCRYPTION: "${MAIL_ENCRYPTION}"
-
-
-    redis:
-        image: redis:alpine
-        ports: ["6379:6379"]
-
-    flow-db:
-        container_name: flow-db
-        image: postgres:13.1
-        ports: ["5432:5432"]
-        environment:
-            POSTGRES_DB : "gaelo-flow"
-            POSTGRES_HOST_AUTH_METHOD: "trust"
-        volumes:
-            - postgres-flow:/var/lib/postgresql/data
-        restart: unless-stopped
-
-    orthanc-flow:
-        image: osimis/orthanc:23.9.0
-        container_name: orthanc-flow
-        restart: unless-stopped
-        ports: ["8042:8042"]
-        volumes: ["orthanc-flow:/var/lib/orthanc/db:Z"]
-        environment:
-            NEURO_PLUGIN_ENABLED: "true"
-            ORTHANC__NAME: "gaelo-flow-orthanc"
-            ORTHANC__DICOM_SERVER_ENABLED: "true"
-            ORTHANC__DICOM_MODALITIES_IN_DATABASE: "true"
-            ORTHANC__ORTHANC_PEERS_IN_DATABASE: "true"
-            DICOM_WEB_PLUGIN_ENABLED: "true"
-            OSIMIS_WEB_VIEWER1_PLUGIN_ENABLED: "true"
-            PYTHON_PLUGIN_ENABLED: "true"
-            TRANSFERS_PLUGIN_ENABLED: "true"
-            WORKLISTS_PLUGIN_ENABLED: "true"
-            ORTHANC__DICOM_WEB__ENABLE: "true"
-            ORTHANC__DICOM_WEB__ROOT: "/dicom-web/"
-            ORTHANC__DICOM_WEB__ENABLEWADO: "true"
-            ORTHANC__DICOM_WEB__WADOROOT: "/wado"
-            ORTHANC__DICOM_WEB__SSL: "true"
-            ORTHANC__DICOM_WEB__STUDIES_METADATA: "MainDicomTags"
-            ORTHANC__DICOM_WEB__SERIES_METADATA: "Full"
-
-volumes:
-        orthanc-flow:
-        postgres-flow:
